Use `inference_engine`'s `encode` and `decode` methods and pass along keywords args. Fixes unwanted special tokens in output.
When the stop sequence is triggered we do a `decode` + `encode` loop which was previously introducing an errant `<|begin_of_text|>` token into the output from the `encode` call. This instructs the tokenizer to avoid adding this special token and potentially others.

We also now use the `encode` and `decode` wrapper on `InferenceEngine` as these do additional threading work which I assume is required for some backends?